==> Data_Cleaning & Feature Engineering/01. Loading_Data&Cleaning.ipynb <==
## Loading Dataframe using Pandas

import pandas as pd

df = pd.read_csv('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/Data/realtor-data.csv')
df

## Cleaning Null Values

df.isna().sum()

"""
Cell generated by Data Wrangler.
"""
from datetime import datetime
import numpy as np
import pandas as pd

def clean_data(df):
    # Drop column: 'street'
    df = df.drop(columns=['street','brokered_by'])
    def months_passed(from_date_str, date_format="%Y-%m-%d"):
        if pd.isna(from_date_str):  # handle NaN/NaT values
            return 0
        # If the column is already datetime, skip parsing
        if isinstance(from_date_str, (datetime, pd.Timestamp)):
            from_date = from_date_str
        else:
            from_date = datetime.strptime(str(from_date_str), date_format)
        today = datetime.today()
        years_diff = today.year - from_date.year
        months_diff = today.month - from_date.month
        total_months = years_diff * 12 + months_diff
        if today.day < from_date.day:
            total_months -= 1
        return total_months
    # Apply to the column
    df['months_since_prev_sold'] = df['prev_sold_date'].apply(months_passed)
    df = df.drop(['prev_sold_date'],axis=1)
    df['bed'] = df['bed'].fillna(df['bed'].mode()[0])
    df['bath'] = df['bath'].fillna(df['bath'].mode()[0])
    df['acre_lot'] = df['acre_lot'].fillna(df['acre_lot'].mean())
    df['house_size'] = df['house_size'].fillna(df['house_size'].mean())
    df = df.dropna()
    return df

df_clean = clean_data(df.copy())
df_clean.head()

df = df_clean.copy()
df.head()

df.isna().sum()

df.shape

df.to_csv('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/Data/Cleaned_Null_Values.csv')

df_sale = df.loc[df['status']!='sold']


df_sale.shape

df_sale.status.unique()

df_sale.to_csv('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/Data/Cleaned_Null_Values+Sale_Data.csv')
