==> Data_Cleaning & Feature Engineering/02. Feature_Engineering.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/Data/Cleaned_Null_Values+Sale_Data.csv')
df

## Feature Engineering

"""
Cell generated by Data Wrangler.
"""
import pickle
import os
from sklearn.preprocessing import LabelEncoder, OneHotEncoder

def clean_data(df):
    os.makedirs('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/pipelines',exist_ok=True)
    le = LabelEncoder()    
    one = OneHotEncoder(sparse_output=False, handle_unknown="ignore")
    status_encoded = one.fit_transform(df[['status']])
    with open('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/pipelines/one_hot_encoding.pkl','wb') as f:
        pickle.dump(one, f)  # save encoder, not data
    status_encoded_df = pd.DataFrame(status_encoded,columns=one.get_feature_names_out(['status']))
    df = pd.concat([df.drop(['status'],axis=1),status_encoded_df],axis=1) 
    df['city'] = le.fit_transform(df['city'] + ", " + df['state'])
    with open('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/pipelines/label_encoding.pkl','wb') as f:
        pickle.dump(le, f)
    df = df.drop(['state'],axis=1)
    df['acre_lot'] = df['acre_lot'] * 43560
    df['total_house_size'] = df['acre_lot'] + df['house_size']
    df['total_rooms'] = df['bed'] + df['bath']
    df['price_per_sqft'] = df['total_house_size'] / df['price']
    df = df.drop(columns=df.columns[df.isin([np.inf, -np.inf]).any()])
    df = df.rename(columns={'price':'cost',
                            'bed':'bedrooms',
                            'bath':'bathrooms',
                            'acre_lot':'sqft_lot',
                            'house_size':'house_lot',
                            'city':'location',
                            'months_since_prev_sold':'last_sold(months)'})
    # Round columns 'cost', 'sqft_lot' and 3 other columns (Number of decimals: 4)
    df = df.round({'cost': 4, 'sqft_lot': 4, 'house_lot': 4, 'total_house_size': 4, 'price_per_sqft': 4})
    return df

df_clean = clean_data(df.copy())
df_clean.head()

df = df_clean.copy()

df

df.isna().sum()

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()

X = pd.DataFrame(
    scaler.fit_transform(df.drop(['cost'], axis=1)),
    columns=scaler.feature_names_in_
)

y = df['cost']

final_df = pd.concat([X,y],axis=1)


final_df

final_df.to_csv('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/Data/Cleaned_Null_Values+Sale_Data+Feature_Engineering.csv')
