==> Data_Cleaning & Feature Engineering/04. Predcition.ipynb <==
import pandas as pd
from tensorflow.keras.models import load_model
import pickle

model = load_model('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/models/best_model.keras',)
with open('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/pipelines/label_encoding.pkl','rb') as f:
    le = pickle.load(f)
with open('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/pipelines/one_hot_encoding.pkl','rb') as f:
    one = pickle.load(f)
with open('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/pipelines/scaler.pkl','rb') as f:
    scaler = pickle.load(f)
with open('D:/Client_Projects/Sir_Malik/Price-Prediction-Model_v2/Data_Cleaning & Feature Engineering/pipelines/scaler2.pkl','rb') as f:
    scaler2 = pickle.load(f)

status = 'for_sale'
bedrooms = 4
bathrooms = 4
sqft_lot = 7200
location = 'Ponce, Puerto Rico'
zip_code = 601.0
house_lot = 5600
last_sold = 5
total_house_size = sqft_lot + house_lot
total_rooms = bedrooms + bathrooms

input_data={'status':status,
            'bedrooms':bedrooms,
            'bathrooms':bathrooms,
            'sqft_lot':sqft_lot,
            'location':location,
            'zip_code':zip_code,
            'house_lot':house_lot,
            'last_sold(months)':last_sold,
            'total_house_size':total_house_size,
            'total_rooms':total_rooms
            }

def predict(input_data,model,le,one,scaler1,scaler2):
    input_df = pd.DataFrame([input_data])   # brackets [] make it a row
    status = one.transform(input_df[['status']])
    status_cols = one.get_feature_names_out(['status'])
    status_df = pd.DataFrame(status, columns=status_cols, index=input_df.index)
    input_df = pd.concat([input_df,status_df],axis=1)
    input_df = input_df.drop(['status'],axis=1)
    input_df['location'] = le.transform(input_df['location'])
    cols = scaler.feature_names_in_
    df_scaled = pd.DataFrame(
        scaler1.transform(input_df[cols]),  # 👈 keep same cols & order
        columns=cols,
        index=input_df.index
    )
    prediction = model.predict(df_scaled)
    return scaler2.inverse_transform(prediction)

print(predict(input_data=input_data,model=model,le=le,one=one,scaler1=scaler,scaler2=scaler2))
